Make changes requested by code review.  Also move copyMml() to MmlNode, since it is useful in order situations, as well.
