+++ components/src/input/tex/extensions/cases/build.json
@@ -1,0 +1,4 @@
+{
+  "component": "input/tex/extensions/cases",
+  "targets": ["input/tex/cases"]
+}
