+++ components/src/input/tex/extensions/cases/cases.js
@@ -1,0 +1,1 @@
+import './lib/cases.js';
