--- components/src/input/tex/extensions/numcases/build.json
@@ -1,4 +1,0 @@
-{
-  "component": "input/tex/extensions/numcases",
-  "targets": ["input/tex/numcases"]
-}
