--- components/src/input/tex/extensions/numcases/numcases.js
@@ -1,1 +1,0 @@
-import './lib/numcases.js';
